Stop ignoring root tsconfig.json file
When specifying files on the commandline, typescript ignores the tsconfig file.
I fixed this by creating multiple tsconfig files in each subproject
which extend the root tsconfig file.

Turning on the strict checks in the root tsconfig.json file exposed a bunch of minor
typescript errors that we hadn't noticed before, so I fixed those too.