Eliminate unneeded dependencies on d3-zoom and d3-geo (re: #263)
We were wrapping d3-mercatorRaw to do projection and inverse,
but we can just copy over the math for this.

We were using d3-zoomIdentity to set up a Transform object
but we can just make an Object with x,y,k properties.

Code in Rapid does not use the d3 methods in either of these,
and we're going to need to own more of the math once we start
rotating the map, per #264