Add `vecRotate`, with code from `geomRotatePoints`
Before `geomRotatePoints` operated on multiple points,
but we need a `vecRotate` to operate on a single point.

This commit also fills out the various rounding functions.
Before we just had `vecFloor`
Now we added `vecRound`, `vecCeil`, `vecTrunc`
I don't know if they'll be used but it's weird for them to be missing.