In tiler tests, use viewport dimensions that make more sense
There are a few tiler tests to test what will happen when the viewport
is along a tile boundary, and they specify a slightly narrower viewport to
avoid fetching tiles outside that viewport.  We should only do this on
tests where we are testing that specific thing, not just copy paste it other
places.