Update Tiler to work on world coordinates
This replaces "pxExtent" which was in a different kind of world coordinates
with "tileExtent" which will be in the world coordinates used everywhere.

The tiler behavior has changed slightly around non-256px tile sizes.
Before we would calculate the tiles in these sizes, now we just adjust the zoom.
I will test to see if this works as expected.